Updates Game Options configuration for Custom Lawset.
This updates the game_options configuration text file A.I lawset default. By default it's configured for 0 which automatically uses Asimov lawset.
I've set it to 1 which will use the laws in Silicon_Laws.txt, this commit alone will do nothing as the default laws for Silicon_Laws.txt are Asimov, however I intend to pull a modified version of Openss13 based laws for Star Trek 13.