Holopad Additions
Adds additional holopads scattered across the Intrepid and the Sovereign. What use is having a holopad in three places if you can't call your friends and tell them to pick up the holopad?
Additionally changes were made to the Akira and Colombia however Git did not appreciate that, and was having issues so I just scrapped the additions on those two for now.