Goodbye Ensigns and Cadets, and all those others dedicated to the Federation
It comes to a time, when we may need to say goodbye to a codebase. As unfortunate as it is, it happens. However people to this day, are still looking back at Openss13, a early opensource branch of space station 13 in Aww. Amazed at what they have to discover. Therefore I find it unreasonable not to say hello and goodbye to the others who come in search of a little Star Trek Space Station 13. A little bit of history never hurt anyone.

Live Long and Prosper everyone.